\name{lukk_small}
\alias{lukk_small}
\title{A dataset based on Lukk \emph{et al}, it contains a subset of 138 samples from more than 5000 in the original article. The gene expression in all the samples is measured using Affymetrix U133A array.
The dataset is a list with 2 slots
\enumerate{
	\item \code{exp} - expression matrix;
	\item \code{annot} - annotation data frame.
}}
\description{
  A dataset based on Lukk \emph{et al}, it contains a
  subset of 138 samples from more than 5000 in the original
  article. The gene expression in all the samples is
  measured using Affymetrix U133A array. The dataset is a
  list with 2 slots \enumerate{ \item \code{exp} -
  expression matrix; \item \code{annot} - annotation data
  frame. }
}
\author{
  Raivo Kolde <rkolde@gmail.com>
}
\references{
  Lukk M, Kapushesky M, Nikkila J, Parkinson H, Goncalves
  A, Huber W, Ukkonen E, Brazma A. "A global map of human
  gene expression." Nat Biotechnology. 2010
  Apr;28(4):322-4.
}
\keyword{data}

