% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{metabolomic_example}
\alias{metabolomic_example}
\title{Example metabolomic dataset}
\description{
\code{metabolomic_example} is a dataset extracted from York \emph{et al} (Metabolights ID:
 MTBLS30), it contains a subset of 120 wild-type samples from 4 tissues: heart, skeletal
 muscle, liver and brain.
}
\details{
The dataset is a list with 2 slots
\enumerate{
	\item \code{data} - metabolite concentration matrix;
	\item \code{annot} - annotation data frame.
}
}
\author{
Raivo Kolde <raivo.kolde@eesti.ee>
}
\references{
York, B., Sagen, J. V., Tsimelzon, A., Louet, J. F., Chopra, A. R., Reineke, E. L.,
 et al. (2013). Research resource: tissue- and pathway-specific metabolomic profiles of the
 steroid receptor coactivator (SRC) family. Mol Endocrinol, 27(2), 366-380.
}
\keyword{data}

