% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{metagenomic_example}
\alias{metagenomic_example}
\title{Example metagenomic dataset}
\description{
\code{metagenomic_example} is a small sample of Human Microbiome Project 16S dataset for
 finding biomarkers characterizing different level of oxygen availability in different
 bodysites. This was downloaded from http://huttenhower.sph.harvard.edu/webfm_send/129.
}
\details{
The dataset is a list with 2 slots
\enumerate{
	\item \code{otu} - otu table for this experiment;
	\item \code{annot} - annotation data frame.
}
}
\author{
Raivo Kolde <raivo.kolde@eesti.ee>
}
\keyword{data}

