\name{is.gosummaries}
\alias{add_to_slot.gosummaries}
\alias{is.gosummaries}
\alias{print.gosummaries}
\title{Functions to interact with gosummaries object}
\usage{
  is.gosummaries(x)

  \method{print}{gosummaries} (x, ...)

  add_to_slot.gosummaries(x, slot, values)
}
\arguments{
  \item{gosummaries}{a gosummaries object}

  \item{\dots}{not used}

  \item{x}{a gosummaries object}

  \item{slot}{the component slot name to be filled (e.g
  Title, Percentage, etc.)}

  \item{values}{list of values where each element
  corresponds to one component}
}
\description{
  Functions to interact with gosummaries object
}
\examples{
#Example1
}
\author{
  Raivo Kolde <rkolde@gmail.com>
}

